+++ app/scripts/controllers/network/index.js
@@ -1,1 +1,1 @@
-export { default, NETWORK_EVENTS } from './network';
+export { default, NETWORK_EVENTS } from './network-controller';
